+++ packages/podspec/src/index.ts
@@ -5,13 +5,18 @@ import type {
   ProofRequestSpec
 } from "./gpc/proof_request.js";
 import { proofRequest } from "./gpc/proof_request.js";
-import type { EntriesOutputType, EntriesSpec } from "./parse/entries.js";
+import type { EntriesSpec } from "./parse/entries.js";
 import { entries } from "./parse/entries.js";
 import type { PodSpec } from "./parse/pod.js";
 import { pod } from "./parse/pod.js";
 import type { EntriesSchema } from "./schemas/entries.js";
 import type { PODSchema } from "./schemas/pod.js";
-import type { InferEntriesType, InferPodType } from "./type_inference.js";
+import type {
+  EntriesOutputType,
+  InferEntriesType,
+  InferJavaScriptEntriesType,
+  InferPodType
+} from "./type_inference.js";
 
 export {
   entries,
@@ -20,6 +25,7 @@ export {
   type EntriesOutputType,
   type EntriesSchema,
   type EntriesSpec,
+  type InferJavaScriptEntriesType,
   type InferEntriesType,
   type InferPodType,
   type ProofConfigPODSchema,